[Data Release] Remove superfluous session checks generating PHP Notices (#3016)
This removes redundant code checking config files and client initialization.

It is redundant because these lines of codes don't affect functionality. Attempts to download files, whether logged in or not, function in the same way whether these lines of code are present or not. Successful download if logged in, 401 status code otherwise.

However, the presence of $client->initialize("../project/config.xml"); causes a PHP Notice when a session already exists everytime a file is downloaded.

This code is copy-pasted widely across LORIS and it is likely worth reviewing to see if it adds any functionality.