[Core] Move most code out of main.php
This moves most of the code out of main.php into two new functions:
displayUserPage (for logged in users) and displayAnonymousPage (for not
logged in users).

This is primarily intended to make it possible for "pages" to be used
as API endpoints without the page frame boilerplate which main.php currently
adds, but also has the side-effect of making the anonymous client template
logic more clearly defined.